Make a unified help handler for !role
This also removes !roles because the same thing can be achieved by
running !role with literally any arguments (or lack thereof).
